imagen de about us actualizada + datos de timeline delegados al servicio
