+++ src/services/DataService.js
@@ -1,4 +1,26 @@
 class DataService {
+	static getTimelineLabels() {
+		const timelineLabels = [
+			{
+				year: 2021,
+				label: "Inicio del emprendimiento ",
+				highlightWords: ["inicio"],
+			},
+			{
+				year: 2022,
+				label: "Compra de la primera impresora",
+				highlightWords: ["impresora"],
+			},
+			{
+				year: 2023,
+				label: "Lanzamiento página web",
+				highlightWords: ["página", "web"],
+			},
+		];
+
+		return timelineLabels;
+	}
+
 	static getQuestions() {
 		const initQA = [
 			{
